a couple more refactors
